requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.7
pydantic-settings==2.9.1
jinja2==3.1.6
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
authlib==1.5.2
httpx==0.28.1
itsdangerous==2.2.0
pypdf==5.4.0
pillow==11.2.1
pytesseract==0.3.13
python-slugify==8.0.4
